+++ src/domain/models/indexer.rs
@@ -40,7 +40,7 @@ pub struct IndexerModel {
     pub process_id: Option<i64>,
     pub target_url: Option<String>,
     pub table_name: Option<String>,
-    pub status_server_port: Option<i16>,
+    pub status_server_port: Option<i32>,
 }
 
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
